% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{model_runner}
\alias{model_runner}
\title{Passes a config to a model, then runs it.}
\usage{
model_runner(path_to_config, path_to_output, verbose)
}
\description{
Passes a config to a model, then runs it.
}
